Modifying for Next.js
