--- app/html/template.html
@@ -1,14 +1,0 @@
-<html>
-  <head>
-    <meta charset="UTF-8" />
-    <meta
-      name="viewport"
-      content="width=device-width, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0"
-    />
-    <title>カフェ塾テラコヤ</title>
-    <!-- <base href="/" /> -->
-  </head>
-  <body>
-    <div id="app"></div>
-  </body>
-</html>
